+++ app/routes/components/Button/changelog.codegen.ts
@@ -5,4 +5,6 @@ export const changelog = ''
   + '- 2025-09-08: Support the prop `ref`.\n'
   + '- 2025-09-09: Support the prop `variant`.\n'
   + '- 2025-09-18: Support the prop `size` (`sm`, `md` and `lg`).\n'
+  + "- 2025-09-23: Update the default button's border color. Add `justify-content:\n"
+  + '  center` to the button styles.\n'
   ;