+++ src/components/Button/index.ts
@@ -1,4 +1,4 @@
-// Button from copy-ui @ 2025-09-18
+// Button from copy-ui @ 2025-09-23
 
 export { Button } from './Button';
 export type { ButtonProps } from './Button';